rogin etc

+++ resources/views/tasks/show.blade.php
@@ -1,8 +1,9 @@
 @extends('layouts.app')
 
 @section('content')
+@if(Auth::check())
+    <h1>{{ Auth::user()->name }}のタスク詳細ページ</h1>
 
-    <h1>id = {{ $task->id }} のタスク詳細ページ</h1>
 
     <table class="table table-bordered">
         <tr>
@@ -20,9 +21,12 @@
     </table>
     
     {!! link_to_route('tasks.edit', 'このタスクを編集', ['id' => $task->id], ['class' => 'btn btn-light']) !!}
+    
+    @if(Auth::id() == $task->user_id)
+        {!! Form::model($task, ['route' => ['tasks.destroy', $task->id], 'method' => 'delete']) !!}
+            {!! Form::submit('削除', ['class' => 'btn btn-danger']) !!}
+        {!! Form::close() !!}
+    @endif
 
-    {!! Form::model($task, ['route' => ['tasks.destroy', $task->id], 'method' => 'delete']) !!}
-        {!! Form::submit('削除', ['class' => 'btn btn-danger']) !!}
-    {!! Form::close() !!}
-
+@endif
 @endsection
